+++ test/diffclj/core_test.clj
@@ -214,5 +214,12 @@
       (is
        (=
         (simplify (deriv '(asin x)))
-        '(/ 1 (sqrt (- 1 (pow x 2)))))))))
+        '(/ 1 (sqrt (- 1 (pow x 2)))))))
+
+
+    (binding [*ns*          (find-ns 'diffclj.core)]
+      (is
+       (=
+        (simplify (deriv '(acos x)))
+        '(/ -1 (sqrt (- 1 (pow x 2)))))))))