+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v0.1.2
     - Added asin 
     - Added acos
+    - Added atan
 
 ------------------